Fixed calculation of IGMRF scale factors
For the purpose of calculating the IGMRF scale factor, the covariance matrix wasn't being calculated correctly (the constraints on the nullspace were wrong).

Further, the generalized variance was not correct in the case of a point constraint. Now the generalized variance is calculated by taking the geometric mean of only the nonzero elements of the diagonal of the covariance matrix. This properly accounts for the fact that the point constraint fixes one element of the IGMRF to be zero (giving it zero variance).