Simplified code to load and process ICCS data
The code to load and process the ICCS data has been simplified. It now uses purrr::map() to read in the various data files into a tibble column and tidyr::unnest() to fill the tibble with the data.

The new code is arguably easier to read (presuming familiarity with purrr and tidyr).